added files from local machine
This is my flashcard app I have been working on in my spare time. The
platform is NodeJS and the framework is ExpressJS. I am connecting to a
local instance on MongoDB as of now for development. Right now the views
are being rendered in Handlebars by Express but this will more than
likely get replaced with React or Angular 2 down the road. Right now you
can create, delete, or edit a deck. You can open a deck but the
functionality to add flashcards is the next step after I secure the all
the current POST request.